requirements.txt
 fastapi==0.83.0
 uvicorn==0.16.0
 psycopg2-binary==2.9.3
 asyncssh==2.12.0
 aiohttp==3.8.3
 elasticsearch==7.17.6
 mobigen-router
 mobigen-service
pydantic~=1.10.2
starlette~=0.19.1
setuptools~=58.1.0
pytz~=2022.6